added tools/create-mandriva-package.sh and updated rest of package creation scripts. Updated site/downloads.html to add sidux also.
git-svn-id: https://svn.code.sf.net/p/socnetv/code@262 fc2d5e9b-b3ce-4657-a66d-b27cdf764933

+++ tools/create-ubuntu-package.sh
@@ -1,24 +1,64 @@
 #!/bin/bash
 
-cd ~/Documents/socnetv/trunk
-
-echo Removing old ubuntu directory
-rm -rf ../ubuntu
 
 
 #CHANGE THIS TO NEW VERSION NUMBERS
-VER=0.50;   
+VER=0.51;   
 echo $VER
+
+echo .
+echo ---------------------------------
+echo    Ubuntu Deb Package Creator
+echo    Copyright Dimitris Kalamaras
+echo	License: GNU GPL v3
+echo ---------------------------------
+echo 
+
+
+
+echo Testing if lintian and devscripts are installed...
+ans=`dpkg-query -l 'lintian*' | grep lintian | awk '{ print $2 }'`
+
+if [ $ans =  "lintian" ]; then 
+	echo OK. Necessary tools are installed; 
+else 
+	echo No lintian or devscripts. Enter sudo password to install them: 
+	sudo apt-get install devscripts pbuilder lintian dput
+fi
+
+
+
+
+cd ~/Documents/socnetv/trunk
+
+echo Removing old ubuntu directory
+if [ -d ../ubuntu ];    then
+        echo Removing old ubuntu directory
+	rm -rf ../ubuntu
+else
+	echo No older ubuntu directory. Continuing...
+fi
+
+
 
 echo .
 echo ---------------------------------
 echo    CLEANING UP COMPILED FILES   
 echo ---------------------------------
 
+./configure > /dev/null 2>&1
 make clean
 rm socnetv 
 
-echo Asking for changelog....
+
+oldfiles=`find . -type f -name *~`;
+for i in $oldfiles; do 
+	echo Removing $i; 
+	rm $i;
+done;
+
+
+echo Asking for changelog entry....
 
 dch -i
 
@@ -28,8 +68,10 @@ echo   COPY FILES TO WORKING DIRS
 echo ---------------------------------
 
 
-find . -not -path "*.svn*" -not -path "*./test-nets*"  -print0  | cpio -pmd0 ../ubuntu/socnetv-$VER
-find . -not -path "*.svn*" -not -path "*./test-nets*"  -print0  | cpio -pmd0 ../ubuntu/socnetv-$VER.orig
+
+find . -not -name "qdevelop-*" -not -name "pajek*" -not -path "*./autom4te.cache*" -not -path "*.svn*" -not -path "*./test-nets*"  -print0  | cpio -pmd0 ../ubuntu/socnetv-$VER
+
+find . -not -name "qdevelop-*" -not -name "pajek*" -not -path "*./autom4te.cache*" -not -path "*.svn*" -not -path "*./test-nets*"  -print0  | cpio -pmd0 ../ubuntu/socnetv-$VER.orig
 
 
 
@@ -43,7 +85,6 @@ echo "Make tarballs? (Y/N)"
 read ans
 if [ $ans = "N" ]; then
         exit;
-
 elif [ $ans = "n" ]; then    
         exit;
 fi
@@ -55,7 +96,7 @@ tar jcfv SocNetV-$VER.tar.bz2 socnetv-$VER/
 
 cd socnetv-$VER/
 
-
+echo 
 echo "Start package creation? (Y/N)"
 read ans
 if [ $ans = "N" ]; then
@@ -67,7 +108,7 @@ fi
 
 echo .
 echo ---------------------------------
-echo    START PACKAGE CREATION       
+echo    START TEST PACKAGE CREATION       
 echo ---------------------------------
 
 debuild 
@@ -75,39 +116,76 @@ debuild
 
 echo .
 echo ---------------------------------
-echo     SOURCE PACKAGE CREATION     
+echo     SOURCE TEST PACKAGE CREATION     
 echo ---------------------------------
 echo .
+
 debuild -S
 
 echo .
 echo ---------------------------------
-echo        TESTING PACKAGE         
+echo        TESTING TEST PACKAGE         
 echo ---------------------------------
 cd ..
+
 lintian -Ivi *.dsc
 
+
+echo Have I build something?
+
+ls *.deb -lh
+
 echo .
-echo ---------------------------------
-echo      INITIAL PACKAGES READY     
-echo ---------------------------------
-echo .
-ls
+if [ -f *.deb ];    then
+        echo test package is ready!
+else
+	echo No DEB...
+fi
+
+echo "Proceed? (Y/N)"
+read ans
+if [ $ans = "N" ]; then
+        exit;
+
+elif [ $ans = "n" ]; then    
+        exit;
+fi
+
+
+
 
 echo .
 echo ---------------------------------
 echo    START FINAL PACKAGE CREATION 
 echo ---------------------------------
 echo .
+
 cd socnetv-$VER/
 debuild -S -sa 
 
-echo .
+echo Check if final DEB has been created...
+echo
 cd ..
-ls 
+ls -lh *.deb 
+echo
+if [ -f *.deb ];    then
+        echo "DEB package is ready! Installing it...";
+	sudo dpkg -i *.deb
+	if [ -f /usr/bin/socnetv ]; then 
+		echo "Package installed OK"; 
+	else 
+		echo "Error Exiting"; 
+		exit; 
+	fi
+else
+	echo "No DEB! Exiting";
+	exit;
+fi
 
 
-echo "Upload package to Launchpad PPA? (Y/N)"
+
+echo . 
+echo "Upload .changes file to Launchpad PPA? (Y/N)";
 read ans
 if [ $ans = "N" ]; then
         exit;
@@ -118,13 +196,44 @@ fi
 
 echo .
 echo ---------------------------------
-echo   UPLOAD FINAL PACKAGE CREATION 
+echo   	UPLOAD FINAL PACKAGE 
 echo ---------------------------------
 echo .
-echo "enter version number, i.e. 0.49-2"
-read VER
+
+VER=`grep urgency socnetv-0.51/debian/changelog | awk '{ print $2 } ' | head -n 1 | sed s/"("// |  sed s/")"//`
+echo New package version: $VER   
+echo "Proceed? (Y/N)"
+read ans
+if [ $ans = "N" ]; then
+        exit;
+
+elif [ $ans = "n" ]; then    
+        exit;
+fi
+
+echo Last exit!
+read ans
 dput ppa socnetv_"$VER"_source.changes
 
+
+
+echo
+echo "Upload package to Sourceforge also? (Y/N)"
+read ans
+if [ $ans = "N" ]; then
+        exit;
+
+elif [ $ans = "n" ]; then    
+        exit;
+fi
+
+echo .
+echo ---------------------------------
+echo   UPLOADING FINAL DEB PACKAGE 
+echo ---------------------------------
+echo .
+
+rsync -avP -e ssh ../ubuntu/*.deb  oxy86@frs.sourceforge.net:uploads/
 
 
 echo --------------------------------
